crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.9
